Add ability to delete an evaluation
