+++ GoalFish/templates/goalfish/eval_detail.html
@@ -12,6 +12,10 @@
     <li>{{evaluation.goal5.description}} {{evaluation.score5}}%</li>
     <li>{{evaluation.goal6.description}} {{evaluation.score6}}%</li>
 </ul>
+<form action="{% url 'goalfish:delete_eval' evaluation.id %}" method="POST">
+        {% csrf_token %}
+    <input type="submit" value="Delete Evaluation">
+</form>
 
 <a href="{% url 'goalfish:student_detail' evaluation.student.id %}">Back to Student Detail</a>
 <a>Back to Student Progress</a>
