+++ GoalFish/urls.py
@@ -19,4 +19,5 @@ urlpatterns = [
     path('grade-filter', views.grade_filter, name='grade_filter'),
     path('weekly-progress-form/<int:student_id>/', views.weekly_progress_form, name='weekly_progress_form'),
     path('weekly-progress-results/<int:student_id>/', views.weekly_progress_results, name='weekly_progress_results'),
+    path('delete-eval/<int:eval_id>/', views.delete_eval, name='delete_eval'),
 ]
