+++ GoalFish/views/__init__.py
@@ -1,3 +1,4 @@
 from .auth import index, register, login_user, user_logout
 from .student import list_students, display_student_form, add_student, student_detail, grade_filter
-from .evaluation import display_eval_form, add_evaluation, eval_detail, weekly_progress_form, weekly_progress_results
+from .evaluation import display_eval_form, add_evaluation, eval_detail, delete_eval
+from .progress import weekly_progress_form, weekly_progress_results
