+++ GoalFish/views/evaluation.py
@@ -64,33 +64,15 @@ def add_evaluation(request, student_id):
 def eval_detail(request, eval_id):
     current_user = request.user
     evaluation = get_object_or_404(Evaluation, pk=eval_id)
+    context = {'evaluation': evaluation}
     template_name = "goalfish/eval_detail.html"
 
-    return render(request, template_name, {'evaluation': evaluation})
+    return render(request, template_name, context)
 
 @login_required(login_url='/login')
-def weekly_progress_form(request, student_id):
-    template_name = "goalfish/weekly_progress_form.html"
-    current_student = get_object_or_404(Student, pk=student_id)
-
-    return render(request, template_name, {'current_student': current_student})
-
-
-@login_required(login_url='/login')
-def weekly_progress_results(request, student_id):
-    template_name = "goalfish/weekly_progress_results.html"
-    current_student = get_object_or_404(Student, pk=student_id)
-    school_week = request.POST["school_week"]
-
-    evaluations = Evaluation.objects.filter(student=current_student, schoolWeek=school_week)
-
-    score1_avg = Evaluation.objects.filter(schoolWeek=school_week, student=current_student).aggregate(Avg('score1'))
-    score2_avg = Evaluation.objects.filter(schoolWeek=school_week, student=current_student).aggregate(Avg('score2'))
-    score3_avg = Evaluation.objects.filter(schoolWeek=school_week, student=current_student).aggregate(Avg('score3'))
-    score4_avg = Evaluation.objects.filter(schoolWeek=school_week, student=current_student).aggregate(Avg('score4'))
-    score5_avg = Evaluation.objects.filter(schoolWeek=school_week, student=current_student).aggregate(Avg('score5'))
-    score6_avg = Evaluation.objects.filter(schoolWeek=school_week, student=current_student).aggregate(Avg('score6'))
-
-
-    return render(request, template_name, {'current_student': current_student, 'evaluations': evaluations, "school_week": school_week, 'score1_avg': score1_avg, 'score2_avg': score2_avg, 'score3_avg': score3_avg, 'score4_avg': score4_avg, 'score5_avg': score5_avg, 'score6_avg': score6_avg })
+def delete_eval(request, eval_id):
+    current_eval = get_object_or_404(Evaluation, pk=eval_id)
+    current_student = current_eval.student.id
+    current_eval.delete()
+    return HttpResponseRedirect(reverse('goalfish:student_detail', args=(current_student, )))
